Move propose lesson component inside requested lessons (resolve #174)
