+++ src/components/RequestedLessonsCard/index.js
@@ -3,12 +3,16 @@ import {Paragraph} from 'egghead-ui'
 import {Text} from 'react-localize'
 import TitleCard from 'components/TitleCard'
 import LessonList from 'components/LessonList'
+import ProposeLesson from './components/ProposeLesson'
 
-export default () => (
+export default ({instructor}) => (
   <TitleCard 
     title={<Text message='requestedLessons.title' />}
     description={<Text message='requestedLessons.description' />}
   >
+    <div className='bb b--gray-secondary'>
+      <ProposeLesson instructor={instructor} />
+    </div>
     <LessonList
       states={['requested']} 
       fallback={
