+++ src/screens/Dashboard/index.js
@@ -4,7 +4,6 @@ import LayoutColumns from 'components/LayoutColumns'
 import InstructorStatsCard from 'components/InstructorStatsCard'
 import InstructorLessonsCard from 'components/InstructorLessonsCard'
 import RequestedLessonsCard from 'components/RequestedLessonsCard'
-import ProposeLessonCard from 'components/ProposeLessonCard'
 import GetPublishedCard from './components/GetPublishedCard'
 import HelpCard from './components/HelpCard'
 import InstructorRevenueCard from './components/InstructorRevenueCard'
@@ -32,10 +31,7 @@ export default ({instructor}) => hasUnlockedPublished(instructor.published_lesso
   : <div>
       <LayoutColumns items={[
         <GetPublishedCard instructor={instructor} />,
-        <div>
-          <ProposeLessonCard instructor={instructor} />
-          <RequestedLessonsCard instructor={instructor} />
-        </div>,
+        <RequestedLessonsCard instructor={instructor} />,
         <HelpCard publishedLessons={instructor.published_lessons} />,
       ]} />
       <InstructorLessonsCard instructor={instructor} />
