+++ src/screens/Lessons/screens/New/index.js
@@ -1,13 +1,8 @@
 import React from 'react'
-import LayoutColumns from 'components/LayoutColumns'
 import RequestedLessonsCard from 'components/RequestedLessonsCard'
-import ProposeLessonCard from 'components/ProposeLessonCard'
 
 export default ({instructor}) => (
   <div>
-    <LayoutColumns items={[
-      <RequestedLessonsCard instructor={instructor} />,
-      <ProposeLessonCard instructor={instructor} />,
-    ]} />
+    <RequestedLessonsCard instructor={instructor} />
   </div>
 )
